Unify cats modules
fixes #28

Credit: @Baccata
