+++ cats/src/test/scala/BaseCatsEffect3Test.scala
