+++ cats/src/test/scala/CatsEffect3TCTests.scala
@@ -1,17 +1,17 @@
 package avocado.tests
 
 import avocado.*
-import avocado.instances.catseffect2.given
+import avocado.instances.cats.given
 
 import scala.concurrent.duration.*
 import cats.Monad
 import cats.syntax.all.*
 import cats.effect.IO
-import cats.effect.IO.given
+import cats.effect.unsafe.implicits.*
 
-class CatsEffect2TCTests extends BaseCatsEffect2Test {
+class CatsEffect3TCTests extends BaseCatsEffect3Test {
 
-  testWithTimeLimit("cats effect 2 comprehension with typeclasses 1", 900) {
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 1", 900) {
     val wait = IO.sleep(500.millis)
     def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
       for {
@@ -21,7 +21,7 @@ class CatsEffect2TCTests extends BaseCatsEffect2Test {
     run[IO](wait)
   }(1)
 
-  testWithTimeLimit("cats effect 2 comprehension with typeclasses 2", 900) {
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 2", 900) {
     val wait = IO.sleep(500.millis)
     def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
       for {
@@ -32,7 +32,7 @@ class CatsEffect2TCTests extends BaseCatsEffect2Test {
     run[IO](wait)
   }(3)
 
-  testWithTimeLimit("cats effect 2 comprehension with typeclasses 3", 1400) {
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 3", 1400) {
     val wait = IO.sleep(500.millis)
     def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
       for {
@@ -45,7 +45,7 @@ class CatsEffect2TCTests extends BaseCatsEffect2Test {
     run[IO](wait)
   }(7)
 
-  testWithTimeLimit("cats effect 2 comprehension with typeclasses 4", 1400) {
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 4", 1400) {
     val wait = IO.sleep(500.millis)
     def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
       for {
@@ -59,7 +59,7 @@ class CatsEffect2TCTests extends BaseCatsEffect2Test {
     run[IO](wait)
   }(7)
 
-  testWithTimeLimit("cats effect 2 comprehension with typeclasses 5", 1400) {
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 5", 1400) {
     val wait = IO.sleep(500.millis)
     def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
       for {
@@ -73,7 +73,7 @@ class CatsEffect2TCTests extends BaseCatsEffect2Test {
     run[IO](wait)
   }(7)
 
-  testWithTimeLimit("cats effect 2 comprehension with typeclasses 6", 1400) {
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 6", 1400) {
     val wait = IO.sleep(500.millis)
     def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
       for {
@@ -89,7 +89,7 @@ class CatsEffect2TCTests extends BaseCatsEffect2Test {
     run[IO](wait)
   }(14)
 
-  testWithTimeLimit("cats effect 2 comprehension with typeclasses 7", 900) {
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 7", 900) {
     val wait = IO.sleep(500.millis)
     def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
       for {
@@ -101,7 +101,7 @@ class CatsEffect2TCTests extends BaseCatsEffect2Test {
     run[IO](wait)
   }(3)
 
-  testWithTimeLimit("cats effect 2 comprehension with typeclasses 8", 900) {
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 8", 900) {
     val wait = IO.sleep(500.millis)
     def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
       for {
@@ -114,7 +114,7 @@ class CatsEffect2TCTests extends BaseCatsEffect2Test {
     run[IO](wait)
   }(3)
 
-  testWithTimeLimit("cats effect 2 comprehension with typeclasses 9", 900) {
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 9", 900) {
     val wait = IO.sleep(500.millis)
     def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
       for {
@@ -128,7 +128,7 @@ class CatsEffect2TCTests extends BaseCatsEffect2Test {
     run[IO](wait)
   }(3)
 
-  testWithTimeLimit("cats effect 2 comprehension with typeclasses 10", 1400) {
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 10", 1400) {
     val wait = IO.sleep(500.millis)
     def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
       for {
@@ -141,7 +141,7 @@ class CatsEffect2TCTests extends BaseCatsEffect2Test {
     run[IO](wait)
   }(3)
 
-  testWithTimeLimit("cats effect 2 comprehension with typeclasses 11", 900) {
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 11", 900) {
     val wait = IO.sleep(500.millis)
     def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
       for {
@@ -155,7 +155,7 @@ class CatsEffect2TCTests extends BaseCatsEffect2Test {
     run[IO](wait)
   }(3)
 
-  testWithTimeLimit("cats effect 2 comprehension with typeclasses 12", 900) {
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 12", 900) {
     val wait = IO.sleep(500.millis)
     def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
       for {
@@ -173,7 +173,7 @@ class CatsEffect2TCTests extends BaseCatsEffect2Test {
     run[IO](wait)
   }(6)
 
-  testWithTimeLimit("cats effect 2 comprehension with typeclasses 13", 900) {
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 13", 900) {
     val wait = IO.sleep(500.millis)
     def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
       for {
@@ -185,7 +185,7 @@ class CatsEffect2TCTests extends BaseCatsEffect2Test {
     run[IO](wait)
   }(6)
 
-  testWithTimeLimit("cats effect 2 comprehension with typeclasses 14", 900) {
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 14", 900) {
     val wait = IO.sleep(500.millis)
     def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
       for {
@@ -197,5 +197,75 @@ class CatsEffect2TCTests extends BaseCatsEffect2Test {
     }
     run[IO](wait)
   }(10)
+
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 15", 1400) {
+    val wait = IO.sleep(500.millis)
+    def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
+      for {
+        a <- wait.map(_ => 1)
+        b = 2
+        c <- wait.map(_ => b + 1)
+        d = 4
+      } yield a + b + c + d
+    }
+    run[IO](wait)
+  }(10)
+
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 16", 1400) {
+    val wait = IO.sleep(500.millis)
+    def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
+      for {
+        a <- wait.map(_ => 1)
+        b = 2
+        c <- wait.map(_ => a + 2)
+        d = 4
+      } yield a + b + c + d
+    }
+    run[IO](wait)
+  }(10)
+
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 17", 1400) {
+    val wait = IO.sleep(500.millis)
+    def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
+      for {
+        a <- wait.map(_ => 1)
+        b <- wait.map(_ => 2)
+        c = a + b
+        d <- wait.map(_ => 4)
+        e <- wait.map(Function.const(5))
+      } yield c + d
+    }
+    run[IO](wait)
+  }(7)
+
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 18", 1400) {
+    val wait = IO.sleep(500.millis)
+    def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
+      for {
+        a <- wait.map(_ => 1)
+        b <- wait.map(_ => 2)
+        c <- Monad[F].pure(a + b)
+        d = 4
+        e <- wait
+      } yield c + d
+    }
+    run[IO](wait)
+  }(7)
+
+  testWithTimeLimit("cats effect 3 comprehension with typeclasses 19", 1400) {
+    val wait = IO.sleep(500.millis)
+    def run[F[_]: Monad: AvocADO](wait: F[Unit]): F[Int] = ado {
+      for {
+        a <- wait.map(_ => 1)
+        b = 2
+        c <- Monad[F].pure(a + b)
+        d <- wait.map(_ => 4)
+        e <- wait
+        f <- wait
+        g <- wait.map(Function.const(7))
+      } yield c + d + g
+    }
+    run[IO](wait)
+  }(14)
 
 }
